refactor(user-api): Separate User interface and UserAltinnAdapter
